packages/jw-datatools/pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "jw-datatools"
version = "0.0.0"
description = "JW Data Utils"
authors = [ "Josh Wycuff <Joshua.Wycuff@turner.com>",]
license = "MIT"
readme = "README.md"
[[tool.poetry.packages]]
include = "jw_datatools"

[tool.poetry.dependencies]
python = ">=3.8,<4"

[tool.poetry.dependencies.jw-errors]
develop = true
path = "../jw-errors"

[tool.poetry.dependencies.jw-types]
develop = true
path = "../jw-types"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
coverage = "^6.5.0"
toml = "^0.10.2"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies.jw-context]
develop = true
path = "../jw-context"
